Fixed build.

+++ src/backend/cpu/CpuBackend.cpp
@@ -309,7 +309,7 @@ rapidjson::Value xmrig::CpuBackend::toJSON(rapidjson::Document &doc) const
     hugepages.PushBack(pages[1], allocator);
 
     out.AddMember("hugepages", hugepages, allocator);
-    out.AddMember("memory",    d_ptr->algo.isValid() ? (ways * d_ptr->algo.memory()) : 0, allocator);
+    out.AddMember("memory",    static_cast<uint64_t>(d_ptr->algo.isValid() ? (ways * d_ptr->algo.memory()) : 0), allocator);
 
     if (d_ptr->threads.empty() || !hashrate()) {
         return out;
